+++ UI/DataChartUI.cs
@@ -32,7 +32,7 @@ namespace MyHome2013
         /// <summary>
         /// Holds the data for each category by month in range being looked at
         /// </summary>
-        private Dictionary<string, Dictionary<DateTime, double>> MonthData { get; set; }
+        private Dictionary<string, Dictionary<DateTime, decimal>> MonthData { get; set; }
 
         #endregion
 
@@ -45,7 +45,7 @@ namespace MyHome2013
         {
             // Intializes the local properties of the form
             this.CategoryNames = new List<string>();
-            this.MonthData = new Dictionary<string, Dictionary<DateTime, double>>();
+            this.MonthData = new Dictionary<string, Dictionary<DateTime, decimal>>();
 
             // Auto generated code for the form
             InitializeComponent();
@@ -168,16 +168,16 @@ namespace MyHome2013
             // Adds the category names as keys
             foreach (string curCategoryName in CategoryNames)
             {
-                MonthData.Add(curCategoryName, new Dictionary<DateTime, double>());
+                MonthData.Add(curCategoryName, new Dictionary<DateTime, decimal>());
             }
 
             // Gets a list with the data of the months in the range
-            Dictionary<DateTime, Dictionary<string, double>> monthData = GetDataForMonthsInRange();
+            Dictionary<DateTime, Dictionary<string, decimal>> monthData = GetDataForMonthsInRange();
 
             int monthRange = MonthsRange();
 
             // Goes over each category getting the total for each month in the range being looked at
-            foreach (KeyValuePair<string, Dictionary<DateTime, double>> curCategoryData in MonthData)
+            foreach (KeyValuePair<string, Dictionary<DateTime, decimal>> curCategoryData in MonthData)
             {
                 DateTime curDate = this.StartDate;
 
@@ -194,9 +194,9 @@ namespace MyHome2013
         /// Gets a dictionary of the totals of each category, per month in the range
         /// </summary>
         /// <returns>A dictionary keyed by date represented, value is the total of each category</returns>
-        private Dictionary<DateTime, Dictionary<string, double>> GetDataForMonthsInRange()
+        private Dictionary<DateTime, Dictionary<string, decimal>> GetDataForMonthsInRange()
         {
-            var monthData = new Dictionary<DateTime, Dictionary<string, double>>();
+            var monthData = new Dictionary<DateTime, Dictionary<string, decimal>>();
 
             DateTime curDate = this.StartDate;
             for (int monthIndex = 0; monthIndex < MonthsRange(); monthIndex++)
